+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+### [upcoming]
+### Added
+- implemented forgot password and reset password feature
 
-### [1.2.0]
+### [1.2.0] 01/10/25
 ### Added
 - Social login feature implemented (Google)
 - OTP based registration system implemented